Fixes size_t

+++ src/config/CryptoNoteConfig.h
@@ -77,7 +77,7 @@ const char     GENESIS_COINBASE_TX_HEX[]                     = "01d80401ff000180
    to import from when the block height cannot be found in the node or the node is offline. */
 const uint64_t GENESIS_BLOCK_TIMESTAMP                       = 1540222033;
 
-const size_t   CRYPTONOTE_REWARD_BLOCKS_WINDOW               = 232.8305;
+const size_t   CRYPTONOTE_REWARD_BLOCKS_WINDOW               = static_cast<size_t>(232.8305);
 const size_t   CRYPTONOTE_BLOCK_GRANTED_FULL_REWARD_ZONE     = 100000; //size of block (bytes) after which reward for block calculated using block size
 const size_t   CRYPTONOTE_BLOCK_GRANTED_FULL_REWARD_ZONE_V2  = 20000;
 const size_t   CRYPTONOTE_BLOCK_GRANTED_FULL_REWARD_ZONE_V1  = 10000;
